protected routes and added user functionality

+++ server/controllers/task.controller.js
@@ -1,7 +1,8 @@
 const Task = require("../models/task.model");
+const User = require("../models/user.model");
 
 const getTasks = async (req, res) => {
-    const tasks = await Task.find();
+    const tasks = await Task.find({ user: req.user.id });
     res.status(200).json(tasks);
 }
 
@@ -15,6 +16,7 @@ const createTask = async (req, res, next) => {
         const { task, isCompleted, completedDate } = req.body;
 
         const newTask = await Task.create({
+            user: req.user.id,
             task,
             isCompleted,
             completedDate
@@ -25,18 +27,26 @@ const createTask = async (req, res, next) => {
         res.status(400);
         next(new Error("Please fill in all details"));
     }
-
-
 }
 
 const updateTask = async (req, res, next) => {
-    const task = await Task.findByIdAndUpdate(req.params.id);
-    for (const data in req.body) {
-        task[data] = req.body[data];
+    const task = await Task.findById(req.params.id);
+    const user = await User.findById(req.user.id);
+
+    if (task.user.toString() !== user.id) {
+        res.status(400);
+        next(new Error("Not authorized"));
     }
+
+    // for (const data in req.body) {
+    //     task[data] = req.body[data];
+    //     console.log(task[data])
+    // }
+
     try {
-        res.status(200).json(task);
-        task.save();
+        const updatedTask = await Task.findByIdAndUpdate(req.params.id, req.body, { new: true });
+        res.status(200).json(updatedTask);
+        // task.save();
     } catch (error) {
         res.status(400);
         next(new Error("This task does not exist on our server"));
@@ -44,9 +54,17 @@ const updateTask = async (req, res, next) => {
 }
 
 const deleteTask = async (req, res, next) => {
-    const task = await Task.findByIdAndDelete(req.params.id);
+    const task = await Task.findById(req.params.id);
+    const user = await User.findById(req.user.id);
+
+    if (task.user.toString() !== user.id) {
+        res.status(400);
+        next(new Error("Not authorized"));
+    }
+
     try {
-        res.status(200).json(task);
+        const deletedTask = await Task.findByIdAndDelete(req.params.id);
+        res.status(200).json(deletedTask);
     } catch (error) {
         res.status(400);
         next(new Error("This task does not exist on our server."));
